% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rds.R
\name{load_rds}
\alias{load_rds}
\title{Read an load the rds files you ask, from dat folder}
\usage{
load_rds(..., folder_path = "./../dat", extension = "rds")
}
\arguments{
\item{...}{Names of the objects you want to load, quoted, without extension.}

\item{folder_path}{Lets you change the directory where data files are stored}

\item{extension}{At the moment it only accepts rds files}
}
\value{

}
\description{
Loads all data objects contained in the project dat folder
}
