% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_path.R
\name{paste_path}
\alias{paste_path}
\title{Reads an excel file exported from S-PRO, containing WIMU data}
\usage{
paste_path(folder_path, file_name)
}
\arguments{
\item{folder_path}{Path to a directory}

\item{file_name}{Name of a file within folder_path}
}
\value{
A character string
}
\description{
It just pastes a path
}
