% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wimu_xls.R
\name{read_wimu_xls}
\alias{read_wimu_xls}
\title{Reads an excel file exported from S-PRO, containing WIMU data}
\usage{
read_wimu_xls(
  path,
  sheet = "Tables",
  session,
  player.var = "player",
  monitor = c("intervalspro"),
  ids = NULL
)
}
\arguments{
\item{path}{Path to the xls file containing WIMU data}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
range. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{session}{Name of the session (character)}

\item{player.var}{Name of the raw column containing ids}

\item{monitor}{Must be intervalspro.}

\item{ids}{Named vector containing the id associated to each wimu, in the
style of c("P101" = "WIMU_1", ...)}
}
\value{
A tibble
}
\description{
This is a specific function for importing player-tracking data captured using
WIMU GPS devices, processed in S-PRO software and exported as a spreadsheet.
read_wimu_xls currently accepts data generated from the Intervals Pro
monitor.
}
